Refactor GridLib to utilize `SpatialInfo` directly instead of `Grid`, update `Size` and `Index` to use `size_t`, adjust temporary file handling in `stitchgrids`, and update related logic and tests.

+++ src/Dem/src/GridLib/ReadDem.cpp
@@ -157,7 +157,7 @@ namespace GridLib
                     rows = b->rows;
                 // DEM files uses columns as the primary dimension, while Grid
                 // uses rows.
-                grid.resize({int64_t(cols), int64_t(rows)});
+                grid.resize({size_t(cols), size_t(rows)});
                 values = grid.values();
             }
             for (int i = 0; i < b->columns; ++i)
